Initialize after the 7th sprint cycle

+++ src/database/views.py
@@ -1,8 +1,10 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth import get_user_model
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
-from django.http import Http404
+from django.http import Http404, HttpResponseRedirect
+from parler.utils import get_active_language_choices
 
 from .forms import ReportForm, RawReportForm, CensorshipForm, RawCensorshipForm
 
@@ -13,8 +15,12 @@ from word.models import Word, Tag
 
 # Create your views here.
 # Database homepage
+@login_required(login_url='login')
 def database_pg(request, *args, **kwargs):
     page = 'database'
+
+    if not request.user.is_superuser:
+        raise Http404
 
     user = get_user_model()
     userlist = user.objects.all().order_by('-date_joined')
@@ -53,8 +59,13 @@ def database_pg(request, *args, **kwargs):
     return render(request, "database/database.html", context)
 
 # Report List
+@login_required(login_url='login')
 def report_list(request):
     page = "list"
+
+    if not request.user.is_superuser:
+        raise Http404
+
     reportlist = Report.objects.filter(Q(option__icontains="Tinjau")).order_by("date")
     donelist = Report.objects.filter(Q(option__icontains="Selesai")).order_by("date")
     censorlist = Censorship.objects.all().order_by('name')
@@ -68,9 +79,13 @@ def report_list(request):
     return render(request, "database/database.html", context)
 
 # View Reports 
+@login_required(login_url='login')
 def report_view(request, srch_id):
     page = 'view'
     report = get_object_or_404(Report, pk=srch_id)
+
+    if not request.user.is_superuser:
+        raise Http404
 
     context = {
         'report': report,
@@ -79,6 +94,7 @@ def report_view(request, srch_id):
     return render(request, "database/report.html", context)
 
 # Report Form
+@login_required(login_url='login')
 def form_rppg(request, pk):
     page = 'form'
 
@@ -108,9 +124,14 @@ def form_rppg(request, pk):
     }
     return render(request, "database/report_form.html", context)
 
+# Report Delete
+@login_required(login_url='login')
 def report_del(request, pk):
     page = 'delete'
     rpt = get_object_or_404(Report, pk=pk)
+
+    if not request.user.is_superuser:
+        raise Http404
 
     if request.method == 'POST':
         rpt.delete()
@@ -122,10 +143,15 @@ def report_del(request, pk):
     }
     return render(request, "database/report_delete.html", context)
 
+# Report Mark as Done
+@login_required(login_url='login')
 def report_done(request, srch_id):
     try:
         obj = Report.objects.get(pk=srch_id)
     except Report.DoesNotExist:
+        raise Http404
+
+    if not request.user.is_superuser:
         raise Http404
     
     print(obj.option)
@@ -133,17 +159,20 @@ def report_done(request, srch_id):
     if obj.option == 'Selesai':
         obj.option = 'Tinjau'
         obj.save()
-        return redirect('database:list')
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
     
     obj.option = 'Selesai'
     obj.save()
-    return redirect('database:list')
+    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
 
 # Censorship CRUD
-
+@login_required(login_url='login')
 def censorship_create(request):
     page = 'censorship'
     form = RawCensorshipForm()
+
+    if not request.user.is_superuser:
+        raise Http404
 
     if request.method == "POST":
         form_data = {
@@ -165,6 +194,7 @@ def censorship_create(request):
     }
     return render(request, 'database/report_form.html', context)
 
+@login_required(login_url='login')
 def censorship_edit(request, csp_id):
     page = 'csp_edit'
     try: 
@@ -172,9 +202,13 @@ def censorship_edit(request, csp_id):
     except Censorship.DoesNotExist:
         raise Http404
 
+    if not request.user.is_superuser:
+        raise Http404
+
     form = CensorshipForm(request.POST or None, instance=obj)
     if form.is_valid():
         form.save()
+        return redirect('database:list')
 
     context = {
         "form": form,
@@ -182,9 +216,13 @@ def censorship_edit(request, csp_id):
     }
     return render(request, "database/report_form.html", context)
 
+@login_required(login_url='login')
 def censorship_delete(request, csp_id):
     page = 'csp_delete'
     censor = get_object_or_404(Censorship, pk=csp_id)
+
+    if not request.user.is_superuser:
+        raise Http404
     
     if request.method == 'POST':
         censor.delete()
